+++ lib/horde/distribution_strategy.ex
@@ -6,7 +6,7 @@ defmodule Horde.DistributionStrategy do
 
   See `Horde.UniformQuorumDistribution` and `Horde.UniformDistribution` for examples.
   """
-  @callback choose_node(identifier :: String.t(), members :: [Horde.Supervisor.Member.t()]) ::
-              {:ok, Horde.Supervisor.Member.t()} | {:error, reason :: String.t()}
-  @callback has_quorum?(members :: [Horde.Supervisor.Member.t()]) :: boolean()
+  @callback choose_node(identifier :: String.t(), members :: [Horde.DynamicSupervisor.Member.t()]) ::
+              {:ok, Horde.DynamicSupervisor.Member.t()} | {:error, reason :: String.t()}
+  @callback has_quorum?(members :: [Horde.DynamicSupervisor.Member.t()]) :: boolean()
 end